Orginize files in Photo input folder. Extract placeholder cell & data providers to separate file.
