+++ ChattoAdditions/Source/Input/Photos/Placeholder/PhotosInputPlaceholderCellProvider.swift
@@ -1,18 +1,18 @@
 /*
  The MIT License (MIT)
-
+ 
  Copyright (c) 2015-present Badoo Trading Limited.
-
+ 
  Permission is hereby granted, free of charge, to any person obtaining a copy
  of this software and associated documentation files (the "Software"), to deal
  in the Software without restriction, including without limitation the rights
  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  copies of the Software, and to permit persons to whom the Software is
  furnished to do so, subject to the following conditions:
-
+ 
  The above copyright notice and this permission notice shall be included in
  all copies or substantial portions of the Software.
-
+ 
  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -24,47 +24,15 @@
 
 import UIKit
 
-class PhotosInputCell: UICollectionViewCell {
-
-    private struct Constants {
-        static let backgroundColor = UIColor(red: 231.0/255.0, green: 236.0/255.0, blue: 242.0/255.0, alpha: 1)
+final class PhotosInputPlaceholderCellProvider: PhotosInputCellProviderProtocol {
+    private let reuseIdentifier = "PhotosPlaceholderCellProvider"
+    private let collectionView: UICollectionView
+    init(collectionView: UICollectionView) {
+        self.collectionView = collectionView
+        self.collectionView.register(PhotosInputPlaceholderCell.self, forCellWithReuseIdentifier: self.reuseIdentifier)
     }
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        self.commonInit()
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)
-        self.commonInit()
-    }
-
-    private var imageView: UIImageView!
-    private func commonInit() {
-        self.clipsToBounds = true
-        self.imageView = UIImageView()
-        self.imageView.contentMode = .scaleAspectFill
-        self.contentView.addSubview(self.imageView)
-        self.contentView.backgroundColor = Constants.backgroundColor
-    }
-
-    override func layoutSubviews() {
-        super.layoutSubviews()
-        self.imageView.frame = self.bounds
-    }
-
-    override func prepareForReuse() {
-        super.prepareForReuse()
-        self.image = nil
-    }
-
-    var image: UIImage? {
-        get {
-            return self.imageView.image
-        }
-        set {
-            self.imageView.image = newValue
-        }
+    
+    func cellForItemAtIndexPath(_ indexPath: IndexPath) -> UICollectionViewCell {
+        return self.collectionView.dequeueReusableCell(withReuseIdentifier: self.reuseIdentifier, for: indexPath)
     }
 }
